Copyright belongs to Peter Van Roy
This library will be a near copy of Peter Van Roy’s original EDCG
library.  In a private email communication, he granted permission for
the library to be released under “any open source license as long as I
am not liable for any use and that proper credit is given if the
package is used”.  The MIT license forbids liability and requires that
credit be given via the mandatory copyright notice, so it meets his
requirements.